Refactor actioner
